+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ## Unreleased
 
+### Features
+
+- Native crash in sample ([#270](https://github.com/getsentry/sentry-unity/pull/270))
 - Cache, background threads and data for UnityEventProcessor ([#268](https://github.com/getsentry/sentry-unity/pull/268))
+
+### Fixes
+
+- Bump Sentry .NET SDK 3.8.3 ([#269](https://github.com/getsentry/sentry-unity/pull/269))
+  - [changelog](https://github.com/getsentry/sentry-dotnet/blob/main/CHANGELOG.md#383)
+  - [diff](https://github.com/getsentry/sentry-dotnet/compare/3.8.2...3.8.3)
 
 ## 0.4.3